+++ CHANGES.md
@@ -48,7 +48,7 @@ Release Notes.
 * Support collection type in dynamic configuration core.
 * Support zookeeper grouped dynamic configurations.
 * Fix NPE when OAP nodes synchronize events with each other in cluster mode.
-* Support configure sampling rate dynamically for service dimension on the backend side.
+* Support configure sampling rate by `configuration module` dynamically and static configuration file `trace-sample-rate-setting.yml` for service dimension on the backend side. Dynamic configurations `agent-analyzer.default.sampleRate` and `agent-analyzer.default.slowTraceSegmentThreshold` are replaced by `agent-analyzer.default.traceSamplingPolicy`. Static configurations `agent-analyzer.default.sampleRate` and `agent-analyzer.default.slowTraceSegmentThreshold` are replaced by `agent-analyzer.default.traceSampleRateSettingFile`.
 
 #### UI